Fix passgo generate to generate proper passwords
There was a problem with `passgo generate` for a long time. It just so happened
that if you requested to generate a very short password, but a very secure pass
then the program would return a password that was much longer than expected.

The reason for this was the program prioritized just getting you a strong
password, rather than getting you a password with your exact length reqs.

This fixes that issue by:
    - Checks if you are requesting an impossible to satisfy password.
    - Short possible to satisfy situations, continue to retry generation
    - Stop returning passwords longer than the user asks for.

Hopefully this makes things work a little better.