start building in module mode
Start building passgo in module mode, which is available in Go 1.11
and newer.

Add go.mod and go.sum files. Since there's already a v2.0.0 tag, start
at v2 instead of v1. This is primarily a command, not a library, so
it's viable not to immediately increase major version by one despite
the pre-go.mod v2.0.0 tag. However, a new v2.1.0 release tag will need
to be created after this change is merged, so that the latest version
includes this change.

Given we're now building in module mode, start using debug.ReadBuildInfo
to determine the version, instead of hardcoding it into the source code.

Remove vendor directory for now; it's no longer necessary and can be
re-added if/when there's a concrete need for it.

Fixes #61